Math.pow

+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/Calculator.java
@@ -10,7 +10,6 @@ public class Calculator {
         return a - b;
     }
 
-    public int square(int a) {
-        return (int) Math.pow(a, 2);
-    }
+    public double square(int a) {
+        return Math.pow(a, 2); }
 }
