+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/ResultChecker.java
@@ -2,7 +2,7 @@ package com.kodilla.basic_assertion;
 
 public class ResultChecker {
 
-    public static boolean assertEquals(int expected, int actual) {
+    public static boolean assertEquals(int expected, double actual) {
         return expected == actual;
     }
 }
